feat(general): add base window manager code
this is the first commit that contains the basis of a window manager that i'm starting to develop.

there are serveral incomplete features, so i've started dividing the tasks into sub-tasks so i can implement everything i want in this wm.
it's also normal for parts of the code that haven't yet been 100% implemented or tested to disappear at this early stage.